before swap  of gasprice
