+++ packages/react-app/src/hooks/index.js
@@ -7,7 +7,6 @@ export { default as useDebounce } from "./Debounce";
 export { default as useEventListener } from "./EventListener";
 export { default as useExchangePrice } from "./ExchangePrice";
 export { default as useExternalContractLoader } from "./ExternalContractLoader";
-export { default as useGasPrice } from "./GasPrice";
 export { default as useLocalStorage } from "./LocalStorage";
 export { default as useLookupAddress } from "./LookupAddress";
 export { default as useNonce } from "./Nonce";
